feat(world): extract ChunkManager and InfiniteStarField from monolithic world.ts
Phase 2 of refactoring plan - decompose large files into focused modules:

**ChunkManager.ts** (1,567 lines extracted)
- Complete chunk-based world generation system
- Celestial object placement and discovery tracking
- Active chunk management and memory optimization
- All chunk coordinate, discovery data, and object interfaces

**InfiniteStarField.ts** (254 lines extracted)
- Parallax background star rendering system
- Multi-layer depth effects with different densities
- Chunk boundary debug visualization
- Memory-efficient region cleanup

**world.ts** (now 23 lines)
- Converted to clean barrel export file
- Re-exports all ChunkManager and InfiniteStarField functionality
- Maintains backward compatibility for existing imports

✅ All 717 tests pass
✅ TypeScript compilation successful
✅ No breaking changes to public APIs

The world system is now properly modularized with each class having focused responsibilities.
